requirements/runtime.txt
boto>=2.20.1
brambox
cachetools>=1.1.6
click
colorama>=0.3.2
dateutils>=0.6.6

delorean
deprecated
docker
flask>=0.10.1
flask-cas
flask-cors

futures_actors
ipython>=5.0.0
lightnet
lockfile>=0.10.2
matplotlib>=3.3.0
networkx>=2.1
numpy>=1.14.2
opencv-python
pandas>=0.19.2
parse>=1.6.6
Pillow>=7.1.0
pyasn1>=0.1.7
pygments>=2.1.3

pynmea2>=1.5.3
pyparsing>=2.1.5

pyqt5>=5.11.2,<5.12
pyzmq>=14.7.0
requests>=2.5.0
scikit-image>=0.12.3
scikit-learn>=0.17.1
scipy>=0.18.0
simplejson>=3.6.5
six>=1.10.0
statsmodels>=0.6.1

torch
torchvision
tornado>=4.2.1
tqdm
ubelt >= 0.8.7
wbia-pydarknet >= 3.0.1
wbia-pyflann >= 3.1.0
wbia-pyhesaff >= 3.0.2
wbia-pyrf >= 3.0.0

wbia-utool >= 3.0.1
wbia-vtool >= 3.2.1
